requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
qdrant-client
python-dotenv
langchain
langgraph
